pyproject.toml
[tool.poetry]
name = "exastro-ita-api-v1"
version = "0.1.0"
description = ""
authors = ["mtakahashiif"]
readme = "README.md"
packages = [
    { include = "exastro_ita_api_v1" },
]

[tool.poetry.dependencies]
python = "^3.10"
lxml = {version = "^5.0.0", optional = true}

[tool.poetry.extras]
webrequest = ["lxml"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
